+++ src/lexicer/lex_utils.rs
@@ -7,6 +7,9 @@ pub fn default_bool() -> bool {
 pub fn default_i32() -> i32 {
     0
 }
+pub fn default_f64() -> f64 {
+    0.0
+}
 // This returns a cloned HashMap
 // because Rust uses the Clone-implementation on HashMap
 pub fn do_clone_map<K: Clone, V: Clone>(data: &HashMap<K,V>) -> HashMap<K, V> {